+++ lib/test_prof/factory_default.rb
@@ -11,14 +11,20 @@ module TestProf
     module DefaultSyntax # :nodoc:
       def create_default(name, *args, &block)
         options = args.extract_options!
-        preserve = options.delete(:preserve_traits)
+        default_options = {}
+        default_options[:preserve_traits] = options.delete(:preserve_traits) if options.key?(:preserve_traits)
+        default_options[:preserve_attributes] = options.delete(:preserve_attributes) if options.key?(:preserve_attributes)
 
         obj = TestProf::FactoryBot.create(name, *args, options, &block)
-        set_factory_default(name, obj, preserve_traits: preserve)
+        set_factory_default(name, obj, **default_options)
       end
 
-      def set_factory_default(name, obj, preserve_traits: nil)
-        FactoryDefault.register(name, obj, preserve_traits: preserve_traits)
+      def set_factory_default(name, obj, preserve_traits: FactoryDefault.preserve_traits, preserve_attributes: FactoryDefault.preserve_attributes)
+        FactoryDefault.register(
+          name, obj,
+          preserve_traits: preserve_traits,
+          preserve_attributes: preserve_attributes
+        )
       end
 
       def skip_factory_default(&block)
@@ -27,7 +33,7 @@ module TestProf
     end
 
     class << self
-      attr_accessor :preserve_traits
+      attr_accessor :preserve_traits, :preserve_attributes
 
       def init
         TestProf::FactoryBot::Syntax::Methods.include DefaultSyntax
@@ -39,24 +45,34 @@ module TestProf
         @enabled = true
         # default is false to retain backward compatibility
         @preserve_traits = false
+        @preserve_attributes = false
       end
 
       def register(name, obj, **options)
-        options[:preserve_traits] = true if FactoryDefault.preserve_traits
         store[name] = {object: obj, **options}
         obj
       end
 
-      def get(name, traits = nil)
+      def get(name, traits = nil, overrides = nil)
         return unless enabled?
 
         record = store[name]
         return unless record
 
         if traits && !traits.empty?
-          return if FactoryDefault.preserve_traits || record[:preserve_traits]
+          return if record[:preserve_traits]
         end
-        record[:object]
+
+        object = record[:object]
+
+        if overrides && !overrides.empty? && record[:preserve_attributes]
+          overrides.each do |name, value|
+            return unless object.respond_to?(name) # rubocop:disable Lint/NonLocalExitFromIterator
+            return if object.public_send(name) != value # rubocop:disable Lint/NonLocalExitFromIterator
+          end
+        end
+
+        object
       end
 
       def remove(name)