+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+- FactoryDefault: Allow creating a default per trait (or set of traits). ([@palkan][])
+
+Now `create_default(:user)` and `create_default(:user, :admin)` would result into two defaults corresponding to the specified traits.
+
 - FactoryDefault: Add stats support. ([@palkan][])
 
 Now you can see how often the default factory values have been used by specifying